streamlit-app/requirements.txt
streamlit
pandas
numpy
matplotlib
altair
requests
python-dotenv
streamlit-aggrid


